accounthandler/requirements.txt
fastapi
pymongo
uvicorn[standard]
python-binance
python-dotenv
pandas